pyproject.toml
[tool.poetry]
name = "sodic"
version = "0.1.0"
description = ""
authors = ["Benjamin Maier <Benjamin.Florian.Maier@web.de>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.4"
matplotlib = "^3.5.2"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-bugbear = "^22.4.25"
flake8-import-order = "^0.18.1"
flake8-bandit = "^3.0.0"
flake8-black = "^0.3.3"
flake8-annotations = "^2.9.0"
flake8-docstrings = "^1.6.0"
pytest = "^7.0"
pytest-cov = "^3.0.0"
safety = "^1.10.3"
mypy = "^0.950"
pre-commit = "^2.17.0"
click = "^8.1.3"
colorama = "^0.4.4"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
format = "ci.format:main"
format_check = "ci.format:main_check"
lint = "ci.lint:main"
test = "ci.test:main"
type_check = "ci.lint:main_mypy"
